Added possibility to dynamically subscribe to mqtt topics
