+++ HABApp/mqtt/__init__.py
@@ -1,1 +1,2 @@
-from .connection import MqttConnection
+from .mqtt_interface import MqttInterface
+from .mqtt_interface import MqttConnection
