+++ HABApp/runtime/__init__.py
@@ -1,2 +1,2 @@
-from .folder_watcher import FileEventTarget
+from .folder_watcher import FileEventTarget, FolderWatcher
 from .runtime import Runtime
